output/dnssim: refactor - move common to separate file
