+++ src/output/dnssim.c
@@ -19,208 +19,16 @@
  */
 #include "output/dnssim.h"
 
+#include "output/dnssim/common.c"
 #include "output/dnssim/udp.c"
 #include "output/dnssim/tcp.c"
 
-uint64_t _now_ms()
-{
-#if HAVE_CLOCK_NANOSLEEP
-    struct timespec ts;
-    uint64_t now_ms;
-    if (clock_gettime(CLOCK_REALTIME, &ts)) {
-        mlfatal("clock_gettime()");
-    }
-    now_ms = ts.tv_sec * 1000;
-    now_ms += ts.tv_nsec / 1000000;
-    return now_ms;
-#else
-    mlfatal("clock_gettime() not available");
-#endif
-}
 
 core_log_t* output_dnssim_log()
 {
     return &_log;
 }
 
-
-/*** request/query ***/
-static void _maybe_free_request(_output_dnssim_request_t* req)
-{
-    if (req->qry == NULL && req->timeout == NULL) {
-        if (req->dnssim->free_after_use) {
-            core_object_payload_free(req->payload);
-        }
-        core_object_dns_free(req->dns_q);
-        free(req);
-    }
-}
-
-static void _close_query(_output_dnssim_query_t* qry)
-{
-    switch(qry->transport) {
-    case OUTPUT_DNSSIM_TRANSPORT_UDP:
-        _close_query_udp((_output_dnssim_query_udp_t*)qry);
-        break;
-    case OUTPUT_DNSSIM_TRANSPORT_TCP:
-        _close_query_tcp((_output_dnssim_query_tcp_t*)qry);
-        break;
-    default:
-        mlfatal("invalid query transport");
-        break;
-    }
-}
-
-static void _close_request(_output_dnssim_request_t* req)
-{
-    if (req == NULL) {
-        return;
-    }
-    if (req->ongoing) {
-        req->ongoing = 0;
-        req->dnssim->ongoing--;
-    }
-    if (req->timeout != NULL) {
-        _close_request_timeout(req->timeout);
-    }
-    // finish any queries in flight
-    _output_dnssim_query_t* qry = req->qry;
-    while (qry != NULL) {
-        _close_query(qry);
-        qry = qry->next;
-    }
-    _maybe_free_request(req);
-}
-
-static void _close_request_timeout_cb(uv_handle_t* handle)
-{
-    _output_dnssim_request_t* req = (_output_dnssim_request_t*)handle->data;
-    free(handle);
-    req->timeout = NULL;
-    _close_request(req);
-}
-
-static void _close_request_timeout(uv_timer_t* handle)
-{
-    _output_dnssim_request_t* req = (_output_dnssim_request_t*)handle->data;
-
-    if (!req->timeout_closing) {
-        req->timeout_closing = 1;
-
-        uint64_t latency = req->ended_at - req->created_at;
-        mlassert(latency <= req->dnssim->timeout_ms, "invalid latency value");
-        req->dnssim->stats_current->latency[latency]++;
-        req->dnssim->stats_sum->latency[latency]++;
-
-        uv_timer_stop(handle);
-        uv_close((uv_handle_t*)handle, _close_request_timeout_cb);
-    }
-}
-
-static void _request_answered(_output_dnssim_request_t* req, core_object_dns_t* msg)
-{
-    req->ended_at = uv_now(&((_output_dnssim_t*)req->dnssim)->loop);
-    if (req->ended_at > (req->created_at + req->dnssim->timeout_ms)) {
-        req->ended_at = req->created_at + req->dnssim->timeout_ms;
-    }
-
-    req->dnssim->stats_sum->answers++;
-    req->dnssim->stats_current->answers++;
-
-    switch(msg->rcode) {
-    case CORE_OBJECT_DNS_RCODE_NOERROR:
-        req->dnssim->stats_sum->rcode_noerror++;
-        req->dnssim->stats_current->rcode_noerror++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_FORMERR:
-        req->dnssim->stats_sum->rcode_formerr++;
-        req->dnssim->stats_current->rcode_formerr++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_SERVFAIL:
-        req->dnssim->stats_sum->rcode_servfail++;
-        req->dnssim->stats_current->rcode_servfail++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_NXDOMAIN:
-        req->dnssim->stats_sum->rcode_nxdomain++;
-        req->dnssim->stats_current->rcode_nxdomain++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_NOTIMP:
-        req->dnssim->stats_sum->rcode_notimp++;
-        req->dnssim->stats_current->rcode_notimp++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_REFUSED:
-        req->dnssim->stats_sum->rcode_refused++;
-        req->dnssim->stats_current->rcode_refused++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_YXDOMAIN:
-        req->dnssim->stats_sum->rcode_yxdomain++;
-        req->dnssim->stats_current->rcode_yxdomain++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_YXRRSET:
-        req->dnssim->stats_sum->rcode_yxrrset++;
-        req->dnssim->stats_current->rcode_yxrrset++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_NXRRSET:
-        req->dnssim->stats_sum->rcode_nxrrset++;
-        req->dnssim->stats_current->rcode_nxrrset++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_NOTAUTH:
-        req->dnssim->stats_sum->rcode_notauth++;
-        req->dnssim->stats_current->rcode_notauth++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_NOTZONE:
-        req->dnssim->stats_sum->rcode_notzone++;
-        req->dnssim->stats_current->rcode_notzone++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_BADVERS:
-        req->dnssim->stats_sum->rcode_badvers++;
-        req->dnssim->stats_current->rcode_badvers++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_BADKEY:
-        req->dnssim->stats_sum->rcode_badkey++;
-        req->dnssim->stats_current->rcode_badkey++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_BADTIME:
-        req->dnssim->stats_sum->rcode_badtime++;
-        req->dnssim->stats_current->rcode_badtime++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_BADMODE:
-        req->dnssim->stats_sum->rcode_badmode++;
-        req->dnssim->stats_current->rcode_badmode++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_BADNAME:
-        req->dnssim->stats_sum->rcode_badname++;
-        req->dnssim->stats_current->rcode_badname++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_BADALG:
-        req->dnssim->stats_sum->rcode_badalg++;
-        req->dnssim->stats_current->rcode_badalg++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_BADTRUNC:
-        req->dnssim->stats_sum->rcode_badtrunc++;
-        req->dnssim->stats_current->rcode_badtrunc++;
-        break;
-    case CORE_OBJECT_DNS_RCODE_BADCOOKIE:
-        req->dnssim->stats_sum->rcode_badcookie++;
-        req->dnssim->stats_current->rcode_badcookie++;
-        break;
-    default:
-        req->dnssim->stats_sum->rcode_other++;
-        req->dnssim->stats_current->rcode_other++;
-    }
-}
-
-
-/* Common network-related stuff. */
-static void _uv_alloc_cb(uv_handle_t* handle, size_t suggested_size, uv_buf_t* buf)
-{
-    mlfatal_oom(buf->base = malloc(suggested_size));
-    buf->len = suggested_size;
-}
-
-
-
-/*** dnssim functions ***/
 output_dnssim_t* output_dnssim_new(size_t max_clients)
 {
     output_dnssim_t* self;
